s3/rpc_server/wsp: store entryid -> docid fwd/reverse mappings
The idea is that a session specific memcache provides a first level
in memory cache that stores copies of the id -> path & path -> id
mappings assoicated with the index specific memcache.

When searching for a value we search in the memory cache first,
if it doesn't exist we search in the index cache, if it doesn't exist
in the index cache we create the mapping (and the reverse one),
if the value already exists in the index memcache  we just populate the
memory cache foward and reverse mappings (and increment a refcount
associated with the records) The refcount is basically a count of
how many unique sessions actually have had a specific id -> path (and
reverse) item returned to it as part of a query result. Even if the
item is no longer used as long as the session (and the associated
client) exists we hold onto the cached id.

When all session caches that reference a id -> path (and reverse) are
gone (e.g. the refcount for that id -> path cache item goes to 0) then
that item is removed from the index cache it belongs to.

We create memory cache(s) for each SMB session we are handling to
minimise memory use, we free those memory caches when all connections
using the associated session are gone.

Pair-Programmed-With: Ralph Boehme <slow@samba.org>
Signed-off-by: Noel Power <noel.power@suse.com>
Reviewed-by: Ralph Boehme <slow@samba.org>